Update script.js
In the end I think the fact it is cas sensitive and order sensitive might be better than the alternative. I could implement a simple regexp and making the array lower case all its strings, but in my opinion it would quickly turn the thing into uselessness as a search bar due to the nature of our alphabet. Since a lot of different letters are present in each activities I'll leave it like this.